requirements.txt
Django>=5.0
gunicorn
whitenoise
requests
markdown
Pygments
tiktoken
openai
pyyaml
